feat(server/src/models/planets.model.js): Added a stream to convert a csv file into a javascript array of objects in order to retrieve the habitable planets only based on the csv file.
Added a function which return a promise that will block the code execution till the event emitter of the stream ends and then resolve the promise.

feat(server/src/server.js):  Import the loadPlanetData function which return a promise, and after creating the server, we add an instruction to await for the promise to be resolved before the server start listenning.